Rename controlled_motor to controller_motor
