+++ Projects/Controller_motor/Controller_motor.ioc
@@ -144,8 +144,8 @@ ProjectManager.MainLocation=Src
 ProjectManager.NoMain=false
 ProjectManager.PreviousToolchain=SW4STM32
 ProjectManager.ProjectBuild=false
-ProjectManager.ProjectFileName=Controlled_motor.ioc
-ProjectManager.ProjectName=Controlled_motor
+ProjectManager.ProjectFileName=Controller_motor.ioc
+ProjectManager.ProjectName=Controller_motor
 ProjectManager.RegisterCallBack=
 ProjectManager.StackSize=0x400
 ProjectManager.TargetToolchain=Other Toolchains (GPDSC)
