+++ Projects/Controller_motor/boards/l0_Controller_motor/l0_Controller_motor.sch
